Don't immediately show an model ID error when changing API providers
